+++ inmanta_plugins/podman/resources/image.py
@@ -27,8 +27,20 @@ import inmanta.resources
 import inmanta_plugins.podman.resources.abc
 
 
+def valid_digest(digest: str, repo_digests: list[str]) -> bool:
+    """
+    Check if the given digest value is a match for one of the repo digests
+    entries.  The format of the repo digests is <repository>:<tag>@<digest>
+
+    :param digest: The digest to validate
+    :param repo_digests: The list of repo digest to compare it to
+    """
+    return digest in [repo_digest.split("@")[-1] for repo_digest in repo_digests]
+
+
 class ImageResource(inmanta_plugins.podman.resources.abc.ResourceABC):
-    pass
+    fields = ("digest",)
+    digest: str | None
 
 
 IR = typing.TypeVar("IR", bound=ImageResource)
@@ -67,6 +79,21 @@ class ImageHandler(
             raise RuntimeError("Failed to inspect image")
 
         return json.loads(stdout)[0]
+
+    def calculate_diff(
+        self,
+        ctx: inmanta.agent.handler.HandlerContext,
+        current: IR,
+        desired: IR,
+    ) -> dict[str, dict[str, typing.Any]]:
+        changes = super().calculate_diff(ctx, current, desired)
+
+        if not changes and desired.digest is None:
+            # When there is no change and there is no desired digest, force the detection
+            # of a change by including the digest in the diff
+            changes["digest"] = {"current": current.digest, "desired": None}
+
+        return changes
 
     def delete_resource(
         self,
@@ -127,6 +154,18 @@ class ImageFromSourceResource(ImageResource):
             options.append(f"--file={entity.file}")
         return options
 
+    @classmethod
+    def get_digest(
+        cls,
+        exporter: inmanta.export.Exporter,
+        entity: inmanta.execute.proxy.DynamicProxy,
+    ) -> None:
+        """
+        The digest can't be known before building the image, so we just
+        return None
+        """
+        return None
+
 
 @inmanta.agent.handler.provider("podman::ImageFromSource", "")
 class ImageFromSourceHandler(ImageHandler[ImageFromSourceResource]):
@@ -141,10 +180,7 @@ class ImageFromSourceHandler(ImageHandler[ImageFromSourceResource]):
             # The image was not found
             raise inmanta.agent.handler.ResourcePurged()
 
-        # Always detect changes, we have no way to know if our image is up to date
-        resource.options = None
-
-        ctx.set("current_digest", existing_image["Digest"])
+        resource.digest = existing_image["Digest"]
 
     def build_image(
         self,
@@ -195,8 +231,11 @@ class ImageFromSourceHandler(ImageHandler[ImageFromSourceResource]):
     ) -> None:
         self.build_image(ctx, resource)
 
-        if ctx.get("current_digest") != self.inspect_image(ctx, resource)["Digest"]:
-            # If digest didn't change, the image was not updated, the rebuild was a noop
+        if not valid_digest(
+            changes["digest"]["current"],
+            self.inspect_image(ctx, resource)["RepoDigests"],
+        ):
+            # If the digest has changed, we have a different image
             ctx.set_updated()
 
 
@@ -224,7 +263,10 @@ class ImageFromRegistryHandler(ImageHandler[ImageFromRegistryResource]):
             # The image was not found
             raise inmanta.agent.handler.ResourcePurged()
 
-        resource.digest = existing_image["Digest"]
+        if not valid_digest(resource.digest, existing_image["RepoDigests"]):
+            # Only detect a different digest if the desired digest (which may be None)
+            # is not part of the image ones
+            resource.digest = existing_image["Digest"]
 
     def pull_image(
         self,
@@ -271,6 +313,9 @@ class ImageFromRegistryHandler(ImageHandler[ImageFromRegistryResource]):
     ) -> None:
         self.pull_image(ctx, resource)
 
-        if changes["digest"]["current"] != self.inspect_image(ctx, resource)["Digest"]:
-            # If digest didn't change, the image was not updated, the pull was a noop
+        if not valid_digest(
+            changes["digest"]["current"],
+            self.inspect_image(ctx, resource)["RepoDigests"],
+        ):
+            # If the digest has changed, we have a different image
             ctx.set_updated()